Move mapper to separate package
