+++ src/main/kotlin/com/chillibits/coronaaid/controller/v1/SymptomController.kt
@@ -2,7 +2,7 @@ package com.chillibits.coronaaid.controller.v1
 
 import com.chillibits.coronaaid.model.dto.SymptomDto
 import com.chillibits.coronaaid.repository.SymptomRepository
-import com.chillibits.coronaaid.shared.toDto
+import com.chillibits.coronaaid.model.mapper.toDto
 import io.swagger.annotations.Api
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.MediaType
